✅ Optimize mock/spy reset
Resetting the value of spies is a costly operation. Avoiding to run this operation all the time will save us lots of CI time.

The helper dealing with declaring cleaning hooks for spies has been adapted to be faster in the context of properties. It used to restore 100 times while it just needed to reset (next run will just re-use the same without needing a full restore).
